// src/api/stetsom/model/getApiLibraryType.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export type GetApiLibraryType =
  (typeof GetApiLibraryType)[keyof typeof GetApiLibraryType];

export const GetApiLibraryType = {
  IMAGE: "IMAGE",
  PDF: "PDF",
  VIDEO: "VIDEO",
  MODEL3D: "MODEL3D",
  MANUAL: "MANUAL",
  CATALOG: "CATALOG",
  CERTIFICATE: "CERTIFICATE",
  IMAGE_PACK: "IMAGE_PACK",
  CATEGORY_ICON: "CATEGORY_ICON",
  OTHER: "OTHER",
} as const;