// src/api/stetsom/model/pageBlockType.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export type PageBlockType = (typeof PageBlockType)[keyof typeof PageBlockType];

export const PageBlockType = {
  TEXT: "TEXT",
  IMAGE: "IMAGE",
  VIDEO: "VIDEO",
  FAQ: "FAQ",
  TIMELINE: "TIMELINE",
  CARDS: "CARDS",
  HTML: "HTML",
  GALLERY: "GALLERY",
} as const;